# server/services/knowledge-server/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponse_DocumentOut_": {
        "properties": {
          "code": {
            "default": 0,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldErrorOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "default": "ok",
            "title": "Message",
            "type": "string"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "title": "ApiResponse[DocumentOut]",
        "type": "object"
      },
...
      "ApiResponse_Page_DocumentOut__": {
        "properties": {
          "code": {
            "default": 0,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Page_DocumentOut_"
              },
              {
                "type": "null"
              }
            ]
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldErrorOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "default": "ok",
            "title": "Message",
            "type": "string"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "title": "ApiResponse[Page[DocumentOut]]",
        "type": "object"
      },
...
      "DocumentOut": {
        "description": "一份文档的样子。",
        "properties": {
          "base_id": {
            "format": "uuid",
            "title": "Base Id",
            "type": "string"
          },
          "byte_size": {
            "title": "Byte Size",
            "type": "integer"
          },
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "failure_reason": {
            "title": "Failure Reason",
            "type": "string"
          },
          "has_raw": {
            "title": "Has Raw",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "ready_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ready At"
          },
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "base_id",
          "source_id",
          "title",
          "media_type",
          "byte_size",
          "status",
          "failure_reason",
          "chunk_count",
          "has_raw",
          "created_at",
          "ready_at"
        ],
        "title": "DocumentOut",
        "type": "object"
      },
      "FieldErrorOut": {
        "description": "字段级校验错误的对外形状。",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "title": "FieldErrorOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_DocumentOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "size",
          "total"
        ],
        "title": "Page[DocumentOut]",
        "type": "object"
      },
...
      "RegisterDocumentIn": {
        "description": "确认直传完成、登记成一份文档。",
        "properties": {
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "filename": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "filename"
        ],
        "title": "RegisterDocumentIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "DigitalTwin Knowledge",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge/documents": {
      "get": {
        "description": "列一个库下的文档，可按摄取状态筛。\n\nArgs: session, _viewer, params, base_id, status_filter。",
        "operationId": "list_all_api_v1_knowledge_documents_get",
        "parameters": [
          {
            "in": "query",
            "name": "base_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Base Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 200,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Page_DocumentOut__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List All",
        "tags": [
          "document"
        ]
      },
      "post": {
        "description": "算哈希、挪进正式键、落一行文档，提交之后投摄取任务。\n\nArgs: session, container, _actor, base_id, body。",
        "operationId": "register_api_v1_knowledge_documents_post",
        "parameters": [
          {
            "in": "query",
            "name": "base_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Base Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDocumentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DocumentOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "确认直传完成",
        "tags": [
          "document"
        ]
      }
    },
    "/api/v1/knowledge/documents/{document_id}": {
      "delete": {
        "description": "删一份文档：先删行，提交之后再清原件。\n\nArgs: session, container, _actor, document_id。",
        "operationId": "drop_api_v1_knowledge_documents__document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Drop",
        "tags": [
          "document"
        ]
      },
      "get": {
        "description": "取一份文档。\n\nArgs: session, _viewer, document_id。",
        "operationId": "get_one_api_v1_knowledge_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DocumentOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get One",
        "tags": [
          "document"
        ]
      }
    },
    "/api/v1/knowledge/documents/{document_id}/figures/{figure_id}": {
      "get": {
        "description": "把一张图的字节吐出去。⚠ 回的是字节不是信封（api-contract 的显式豁免）。\n\nArgs: session, container, _viewer, document_id, figure_id。",
        "operationId": "read_figure_api_v1_knowledge_documents__document_id__figures__figure_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "figure_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Figure Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "取一张解析出来的图",
        "tags": [
          "document"
        ]
      }
    },
    "/api/v1/knowledge/documents/{document_id}/preview": {
      "get": {
        "description": "优先吐 DOCX 的 PDF 派生物，缺席时回原件兼容预览。\n\nArgs: session, container, _viewer, document_id。",
        "operationId": "read_preview_api_v1_knowledge_documents__document_id__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "PDF 派生预览，未生成时为原 DOCX"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "取 DOCX 原件的页面预览",
        "tags": [
          "document"
        ]
      }
    },
    "/api/v1/knowledge/documents/{document_id}/raw": {
      "get": {
        "description": "把一份文档的原件吐出去，供页面里预览与下载。\n\n⚠ 回的是字节不是信封（与图那条同一份 api-contract 显式豁免）。\n\nArgs: session, container, _viewer, document_id。",
        "operationId": "read_raw_api_v1_knowledge_documents__document_id__raw_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "取原件的字节",
        "tags": [
          "document"
        ]
      }
    },
    "/api/v1/knowledge/documents/{document_id}:reparse": {
      "post": {
        "description": "把一份文档退回待处理并重新排队。\n\n⚠ 这是这条链路上**唯一**的重试入口，而且它由人按：一份解不动的文档\n自动重试一万次也解不动，只会把 worker 占满。\n\nArgs: session, container, _actor, document_id。",
        "operationId": "reparse_api_v1_knowledge_documents__document_id__reparse_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DocumentOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "重新解析",
        "tags": [
          "document"
        ]
      }
    },
...
  }
}